+++ xup/xup.c
@@ -212,7 +212,8 @@ lib_mod_connect(struct mod *mod)
         if (trans_connect(mod->trans, mod->ip, con_port, 3000) == 0)
         {
             LLOGLN(0, ("lib_mod_connect: connected to Xserver "
-                   "(Xorg or X11rdp) sck %ld", (long) (mod->trans->sck)));
+                   "(Xorg or X11rdp) sck %lld",
+                   (long long) (mod->trans->sck)));
             error = 0;
         }
